Add number converter for response data
The API returns all quantitative data as strings, presumably in order to
preserve precision. String numbers are a pain to work with, though, so
here's a converter to use that'll make all of those strings into proper
Numeric instances (mainly BigDecimal in order to continue to preserve
precision).

This still needs to be _used_ on responses we get. Still working on that
public-interface part of it.